requirements.txt
clinner==1.12.3
colorlog==3.2.0
DateTime==4.3
Django==2.2.6
django-elasticsearch-dsl==7.0.0
django-elasticsearch-dsl-drf==0.20.3
django-nine==0.2.2
djangorestframework==3.10.3
elasticsearch==7.0.5
elasticsearch-dsl==7.0.0
git+https://github.com/RockefellerArchiveCenter/asterism@v0.1#egg=asterism
git+https://github.com/RockefellerArchiveCenter/rac_es#egg=rac_es
gitdb2==2.0.6
GitPython==3.0.3
health-check==3.4.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
shortuuid==0.5.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
zope.interface==4.6.0
